[#1375] Revise public API
move AgreementsResource and Factory to internal package
