+++ bundles/org.eclipse.passage.lic.agreements.model/src/org/eclipse/passage/lic/internal/agreements/model/util/AgreementsResourceFactoryImpl.java
@@ -1,0 +1,26 @@
+/*******************************************************************************
+ * Copyright (c) 2018, 2024 ArSysOp
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * https://www.eclipse.org/legal/epl-2.0/.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     ArSysOp - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.passage.lic.internal.agreements.model.util;
+
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.ecore.resource.Resource;
+import org.eclipse.emf.ecore.resource.impl.ResourceFactoryImpl;
+
+public class AgreementsResourceFactoryImpl extends ResourceFactoryImpl {
+
+	@Override
+	public Resource createResource(URI uri) {
+		return new AgreementsResourceImpl(uri);
+	}
+
+}
