+++ bundles/org.eclipse.passage.loc.agreements.core/src/org/eclipse/passage/loc/internal/agreements/core/AgreementDomainRegistry.java
@@ -29,7 +29,7 @@ import org.eclipse.osgi.util.NLS;
 import org.eclipse.passage.lic.agreements.model.api.Agreement;
 import org.eclipse.passage.lic.agreements.model.api.AgreementGroup;
 import org.eclipse.passage.lic.agreements.model.meta.AgreementsPackage;
-import org.eclipse.passage.lic.agreements.model.util.AgreementsResourceImpl;
+import org.eclipse.passage.lic.internal.agreements.model.util.AgreementsResourceImpl;
 import org.eclipse.passage.lic.internal.equinox.events.EquinoxEvent;
 import org.eclipse.passage.loc.internal.agreements.AgreementRegistry;
 import org.eclipse.passage.loc.internal.agreements.AgreementRegistryEvents;
